Pass proper data format to ldap.modlist
Seems like data passed to ldap.modlist has been formatted in a way, that
had not been covered by python-ldap tests. ({str: str}) Somewhere around
python-ldap 3.0 this got broken, so that in 3.1.0 result contains list
of letters of input values.

This patches most uses of modlist, so that {str: [str]} is used instead.